feat: Upgrade x86_64 crate to version 0.15

+++ pkg/kernel/src/interrupt/clock.rs
@@ -6,7 +6,7 @@ use core::sync::atomic::{AtomicU64, Ordering};
 use x86_64::structures::idt::{InterruptDescriptorTable, InterruptStackFrame};
 
 pub unsafe fn register_idt(idt: &mut InterruptDescriptorTable) {
-    idt[Interrupts::IrqBase as usize + Irq::Timer as usize]
+    idt[Interrupts::IrqBase as u8 + Irq::Timer as u8]
         .set_handler_fn(clock_handler)
         .set_stack_index(gdt::TIMER_IST_INDEX);
 }
