+++ pkg/kernel/src/interrupt/serial.rs
@@ -4,7 +4,7 @@ use super::consts::*;
 use x86_64::structures::idt::{InterruptDescriptorTable, InterruptStackFrame};
 
 pub unsafe fn register_idt(idt: &mut InterruptDescriptorTable) {
-    idt[Interrupts::IrqBase as usize + Irq::Serial0 as usize].set_handler_fn(serial_handler);
+    idt[Interrupts::IrqBase as u8 + Irq::Serial0 as u8].set_handler_fn(serial_handler);
 }
 
 pub extern "x86-interrupt" fn serial_handler(_st: InterruptStackFrame) {
