+++ pkg/kernel/src/interrupt/syscall/mod.rs
@@ -18,7 +18,7 @@ pub unsafe fn register_idt(idt: &mut InterruptDescriptorTable) {
     // register syscall handler to IDT
     // - standalone syscall stack
     // - ring 3
-    idt[consts::Interrupts::Syscall as usize]
+    idt[consts::Interrupts::Syscall as u8]
         .set_handler_fn(syscall_handler)
         .set_stack_index(gdt::SYSCALL_IST_INDEX)
         .set_privilege_level(x86_64::PrivilegeLevel::Ring3);
